refactor(ast-extractors): pull out the es node identification functions into estree-helpers.js
